refactor core.rb file, split into 3 files, extract methods from parse_spec
